+++ libffmpegxx/avformat/MediaInfoFactory.cpp
@@ -40,38 +40,15 @@ StreamType getTypeFromCodecId(AVMediaType codecType) {
 VideoInfo buildVideoStreamProperties(AVStream *stream) {
   VideoInfo info;
 
-  info.averageFramerate = av_q2d(stream->avg_frame_rate);
+  info.averageFramerate = {stream->avg_frame_rate.num,
+                           stream->avg_frame_rate.den};
   info.frameCount = stream->nb_frames;
-  info.format = static_cast<AVPixelFormat>(stream->codecpar->format);
-  info.height = stream->codecpar->height;
-  info.width = stream->codecpar->width;
-  info.sampleAspectRatio = {stream->codecpar->sample_aspect_ratio.num,
-                            stream->codecpar->sample_aspect_ratio.den};
-  info.chromaLocation = stream->codecpar->chroma_location;
-  info.colorPrimaries = stream->codecpar->color_primaries;
-  info.colorRange = stream->codecpar->color_range;
-  info.colorSpace = stream->codecpar->color_space;
-  info.colorTrc = stream->codecpar->color_trc;
-  info.fieldOrder = stream->codecpar->field_order;
-  info.videoDelay = stream->codecpar->video_delay;
 
   return info;
 }
 
 AudioInfo buildAudioStreamProperties(AVStream *stream) {
   AudioInfo info;
-
-  info.format = static_cast<AVSampleFormat>(stream->codecpar->format);
-  info.channelCount = stream->codecpar->channels;
-  info.channelLayout = stream->codecpar->channel_layout;
-  info.frameSize = stream->codecpar->frame_size;
-  info.sampleRate = stream->codecpar->sample_rate;
-  info.bitsPerCodedSample = stream->codecpar->bits_per_coded_sample;
-  info.bitsPerRawSample = stream->codecpar->bits_per_raw_sample;
-  info.blockAlign = stream->codecpar->block_align;
-  info.initialPadding = stream->codecpar->initial_padding;
-  info.seekPreroll = stream->codecpar->seek_preroll;
-  info.trailingPadding = stream->codecpar->trailing_padding;
 
   return info;
 }
@@ -119,15 +96,7 @@ StreamInfo buildStreamInfo(AVStream *stream) {
   info.startTime = time::Seconds{
       av_rescale_q(stream->start_time, stream->time_base, {1, 1})};
   info.timebase = time::Timebase(stream->time_base.num, stream->time_base.den);
-  info.bitrate = stream->codecpar->bit_rate;
-  info.level = stream->codecpar->level;
-  info.profile = stream->codecpar->profile;
-  info.codecTag = stream->codecpar->codec_tag;
-  info.codecType = stream->codecpar->codec_type;
-  info.extraDataSize = stream->codecpar->extradata_size;
-  info.extraData = std::vector<uint8_t>(stream->codecpar->extradata,
-                                        stream->codecpar->extradata +
-                                            stream->codecpar->extradata_size);
+  info.codecPar = avcodec::AVCodecPar(stream->codecpar);
   info.metadata = utils::fromAVDictionary(stream->metadata);
 
   return info;