+++ libffmpegxx/include/avcodec/DecoderImpl.h
@@ -11,9 +11,11 @@ public:
   ~DecoderImpl();
 
   int decode(IAVPacket *packet, avutil::IAVFrame *frame) override;
+  int flush(std::vector<avutil::IAVFrame *> &flushedFrames) override;
 
 private:
   AVCodecContext *m_codecCtx{nullptr};
+  time::Timebase m_tb;
 };
 }; // namespace avcodec
 }; // namespace libffmpegxx