+++ libffmpegxx/include/utils/exception.h
@@ -2,9 +2,11 @@
 
 #include "utils/LoggerApi.h"
 
-#define LOG_FATAL_FFMPEG_ERR(description, errCode)                     \
+#include <sstream>
+
+#define LOG_FATAL_FFMPEG_ERR(description, errCode)                             \
   std::stringstream sstream;                                                   \
   sstream << description;                                                      \
-  sstream << "Error code " << std::to_string(errCode) << ": ";                 \
+  sstream << " Error code " << std::to_string(errCode) << ": ";                \
   sstream << utils::Logger::avErrorToStr(errCode);                             \
   LOG_FATAL(sstream.str());